abstracted metric cache class and added pytest unit tests with clearera ssertions
